+++ src/lib/viewers/controls/media/PlayPauseToggle.tsx
@@ -30,15 +30,30 @@ export default function PlayPauseToggle({ isPlaying, onPlayPause = noop, movePla
 
     return (
         <>
-            <MediaToggle className="bp-PlayPauseToggle" onClick={moveBackward} title={skipBackwardTitle}>
+            <MediaToggle
+                className="bp-PlayPauseToggle"
+                data-resin-target="skipBackward"
+                onClick={moveBackward}
+                title={skipBackwardTitle}
+            >
                 <IconBack24 />
             </MediaToggle>
 
-            <MediaToggle className="bp-PlayPauseToggle" onClick={() => onPlayPause(!isPlaying)} title={title}>
+            <MediaToggle
+                className="bp-PlayPauseToggle"
+                data-resin-target="playPause"
+                onClick={() => onPlayPause(!isPlaying)}
+                title={title}
+            >
                 <PlayPauseIcon />
             </MediaToggle>
 
-            <MediaToggle className="bp-PlayPauseToggle" onClick={moveForward} title={skipForwardTitle}>
+            <MediaToggle
+                className="bp-PlayPauseToggle"
+                data-resin-target="skipForward"
+                onClick={moveForward}
+                title={skipForwardTitle}
+            >
                 <IconForward24 />
             </MediaToggle>
         </>